Fix lint issues: remove unused code and reduce test duplication
- Remove 7 unused variables, functions, and types
- Add generic test helper functions to eliminate code duplication
- Refactor validation tests to use shared helpers
- Reduce dupl issues from 10 to 4 (60% improvement)

+++ pkg/sbdprotocol/message_test.go
@@ -96,7 +96,7 @@ func TestNewFence(t *testing.T) {
 	nodeID := uint16(42)
 	targetNodeID := uint16(99)
 	sequence := uint64(456)
-	reason := uint8(FENCE_REASON_HEARTBEAT_TIMEOUT)
+	reason := FENCE_REASON_HEARTBEAT_TIMEOUT
 
 	before := time.Now().UnixNano()
 	msg := NewFence(nodeID, targetNodeID, sequence, reason)
